apps/backend/requirements.txt
boto3>=1.34.0
gitpython>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
mangum>=0.17.0
pyjwt>=2.8.0
httpx>=0.26.0
cryptography>=42.0.0
requests>=2.31.0
